Move products and prices from stripe util
